Optionally verify pgp signatures
I could not get `gkeys verify` to work and gkeys-gpg does not work with git
(#595850), so call gpg directly.

Given that keys change more seldomly than releases, if overlay users use this
as "trust on first use", there are fewer opportunities for interception.

But hopefully everyone will verify the keys out of band...