www-client/torbrowser: Fix musl build
Gentoo uses target triples with `pc` or `gentoo` instead of `unknown` as
a vendor. However, such triples are not understood by Cargo. The
patchset tarball already contains a patch for overriding the target, but
it needs to be defined in the `RUST_TARGET` environment variable.

Disable jemalloc for now, since Mozilla's jemalloc doesn't build on
musl.

All these changes are ported from firefox-128 ebuild.